Compatibility & .FSE Utility
##Overview
Purpose of this utility is just to simplify the setting of compatibility flags for exe-files - those from "Compatibility" tab in Properties dialog of exe-files.

##Features:
1. Searching for exe-files in folder (and its subfolders) which is user specified.
2. Flags available to edit are:
- Run this program as an administrator
- Disable full-screen optimizations
- Run this program in compatibility mode for {versions of Windows}.
3. More than one cell in checkbox columns can be selected by standard means with keyboard (press Up or Down holding Shift) or with a mouse (click any cells holding Ctrl or click first and last cells in range holding Shift) and then toggled by Space.
4. Jump to file location by double click on the cell with exe-file path.
5. Preserving last searched folder between utility launches.
6. All compatibility flags are saved for the current user only.
Update: 7. Searching in the registry (for current user) for already configured exe-files (button "Check registry").

When done with editing just need to click the button "Apply".

To start a search just type folder in the edit box at toolbar - it will give suggestions just like address bar in explorer, and then click button "Search files".


##Requirements
Utility is built with .Net Framework 4.5
Elevated execution is not needed but I only tried it with an administrator account.