Each update the damn HPET question via eMail, so let's end it
HPET
=====
In the case of Windows 8 and 10, you need to enable HPET when using Ryzen Master, because in AMD CPUs the TSC is tied to the base clock (P0 state). When you change the base clock while the OS is running you break the UI compositor (Aero), so you get a black screen. Windows 7 can run without hardware accelerated Aero, so it can fall back to that. HPET itself runs independently from TSC, so you can enable HPET, use Ryzen Master to test the desired overclock, disable the HPET, and finally set up the OC parameters in BIOS. That way you don't lose performance to HPET.

Why some systems (Windows 10) have HPET enabled and why others not?!
This is depending on the driver and if the OS detects it via the BIOS/UEFI use bcdedit /enum to determinate the status BEFORE you touch anything. If it's not present then just don't touch something if it's disabled leave it that way. If it's enabled you can "play" with it and see if that helps. 


Yes Intel and AMD have different opinions and implementations here, see here: https://www.anandtech.com/show/12678/a-timely-discovery-examining-amd-2nd-gen-ryzen-results/3