ASRock Timing Configuration Utility 4.0.4 (mirror)
I was not able to find an official landing page, so I mirror the latest (?) version of the utility here. If someone find a real landing and download site (I only found v 4.0.3 on the official page) feel free to make a pull request. I do not claim to have any copyright on this tool and provide it "as-is" as a mirror.

The Timing Configuration Utility does not only work with ASRock systems - basically spoken, there are no checks within the program - which means you can OC your RAM with this utility how you want. You can adjust the timings directly under Windows.