Research: Wireless Mouse Click Latency Analysis Using Breadboard & USB Analyzers
Logitech G900 and G403 Vs. Razer DeathAdder Chroma was tested.

Overall result:
* Logitech: 3~6ms (wireless)
* Razer 9~12ms (cable)

My explanation from 1997 (when this topic started, the hardware forum does not exist anymore):
"Signals are transmitted by the speed of light. Cables (CU) using ~67% C. On both sides the signal needs to be "translated" and processed. So in theory wireless should be equal or even faster (depending on the translation process) compared to cable.

This basically debunks the myth that "wireless" has a higher latency. Does that matter at the end? No! Why not? Simply because the human response is much slower, Steve also mentioned that in the article. 

However, it's an interesting article, and you can see what hardware they used to test the "click latency".