Added my EAPO settings
How to use it?
- Install EAPO, enable it and restart Windows (or restart explorer + "sound" via services.msc)
- Copy the microphone.peace file into the EAPO folder and restart (if running) the .exe
- You can (if you like to) use my config.txt (just override/backup your current/default config.txt) in the EAPO folder. This basically sets the sliders, otherwise use the screenshots I provided to setup the sliders and the frequencies. 

About:
I use a Rode NT (USB mic) so you might want to change the pre-amp (since there is no "mic boost" anymore since Windows 10 -> only with "special R2" Realtek drivers (there are not worth it, use pre-amp!) 3-5 is typically enough if you set the Windows "ouput" slider to max 85% for the actual recording.

The given example is nothing but this, an example! It's a matter of taste of course, some people want to sound "boomy" (voice with bass) others want to get rid of specific stuff, however you can just go ahead use my example and increase the bass if you want to. I'm personally a really thin person and if I sit, I sound like "Hulk Hogan" with lots of bass in my voice, that's why I actually try to get rid of it, since I prefer a "clear" voice over a "boomy" one. The low frequencies as well as 20+k Hz frequencies are cut out because you actually want to get rid of "keyboard noise" and the frequencies which no human ear is going to hear. Of course, it DOES NOT magically "filter" keyboard or mouse clicks once the "microphone is open" but you will not notice it if it's a bit away from your mouth and your PC fans. If you want additional filtering you could use low-pass and high-pass filters, but typically I would say EAPO does a better job than OBS Studio own plugins because those a) adding latency b) you simply sound "muddy" or it simply "changes" too much (as a negative side-effect), this is not the case with EAPO.  OBS own VST's are more in-depth explained in the first tutorial link, you also could combine the findings from the first tutorial and combine it with the stuff you learn from the second tutorial but again I say you not need that unless you have a 10 dollar mic. 

The example will work even if you do not use a Rode NT, you just need to give EAPO your device ID (select your device from the right-side of the GUI) and then you will notice the change. 

I do not provide a speaker or headset .peace file because this depends a lot of things, mostly own preferences and it's easier to create your own .peace file for that, just use one of the pre-made (integrated ones), change it and save it under a different name. 

This is explicitly for USB / 3.5mm, in case you use a more "professional" Mic like an Mic with hardware mixer then you might not need it unless your hardware mixer does not offer all the functions which EAPO supports or in case you have multiple sound devices/drivers.