Updated the page file myth
Open "Virtual Memory":
Win + R; "systempropertiesperformance"

Different names:
The official documentation, which was updated several weeks ago, calls it a "page file" The Windows dialog box, which dates back more than 20 years, calls it a "paging file".

Official documentation:
https://docs.microsoft.com/en-us/windows/client-management/introduction-page-file#support-for-system-crash-dumps

What to mention:
Some people argue that you can reclaim disk space by eliminating the page file completely e.g. on systems with >= 32GB, others recommend setting it to a fixed size, so that you don't experience a performance hit when it automatically resizes itself. In my tests, even if Windows resize the page file automatically during a game session, the process is so fast that you won't notice it, we are talking about 1-2 ms (avg).

In my tests, with several games setting the page file to a fixed file NEVER "fixed" frame stutter or FPS, this is a claim without anything and I do not understand why this is still written e.g. here: https://www.pcgamingwiki.com/wiki/Resident_Evil_2_(2019). I checked it and it did nothing. This is an "engine" related stutter and has nothing to do with Windows issue or the page file on its own. The initial measurement time is done once the game is loaded so BEFORE you are in the game session.