2 years after: Benchmarking Spectre/Meltdown mitigations
My comment:
- No tests done with all mitigations enabled!
- Zen 1 is almost identical here in the real-world tests affected to Zen+, which included some optimizations in RAM access and cache latency, but absolutely no core redesign.
- A new architecture that is built from scratch to avoid this family of vulnerabilities will take several years to develop and even then it's unclear if they are faster or not (clock by clock wise compared).

Nutshell:
Based on the numbers you see that each "new" CPU generation (refresh) is less affected overall spoken. The numbers are so low (and there is always margin of errors) that in-game and application wise it should not matter that much to leave the migrations enabled. It's unclear which migrations are most critical since the test did not test all migration options.

> Core i7 5960X saw 86% the performance out-of-the-box compared to the unmitigated performance.
> Core i7 8700K about 84% the unmitigated performance
> Core i5 9400F at 86%
> 86%, the Core i9 9900K at 88%
> Intel Core i9 10980XE at 97%
> E3-1280 v5 was at 77%
> Xeon E3-1275 v6 at 79%
> Xeon Platinum 8280 Cascade Lake came in at 95%

Verdict:
Benchmarks are slowed down by around ~10-12%. The Phoronix Test Suite was used (real-word) workloads tested. Some of the HW fixes seem to help without hurting the performance, see the "ctx_clock" test for instance. That been said, some migrations are more "performance critical" than others, which means to overall saying you "need" to disable all migrations in the OS is definitely debunked now. Robbing the user of the opportunity to get some performance back by switching the mitigations off seems rather idiotic.