Sound: Mention Daniel_K & PAX modded sound drivers
Nutshell:
-> ASUS/Creative/Realtek modded drivers/packages
-> Daniel_K seems "missing"
-> PAX is still active, lots of his work is based on Daniel_K's packages and tweaks.

Modded drivers/packages are a matter of taste, back in Windows 7 times they had its benefits because Creative and Co. decided (due to license reasons) to strip away some features like Dolby which you could unlock via a modded package. There are some other benefits but's a matter of taste and in Windows 10 times it became a little bit redundant, so I suggest you read the official pages, create a restore point and decide to install/test the drivers based on the description if you need additional "sound tweaks".

Sound itself is a huge and interesting topics, there are different opinions on "what you need" and what you "better choose", but I believe everyone has to test it for himself to come to a conclusion. I used (back in my Creative PCIe days), the modded package to "fix" my bass issue and it worked very well.