Sound: Mention "The Ultimate Realtek HD Audio Driver Mod for Windows 10" thread
Warning/Info:
-> Not all Realtek nor Windows 10 versions are working or have been tested.
-> The thread is still active and in case your Realtek chip is not supported you can post your Device ID to possible get support for your chip. If you do not know how to obtain the Device ID or if you're unsure how to identify your Chip (e.g. via AIDA/hwinfo) you can follow the guidance given over here: https://www.tenforums.com/sound-audio/125936-determine-hda-realtek-driver-needed-your-audio.html#post1561988
-> Using more sound effects will affect latency as well as CPU usage. Currently, there is no mod to configure the buffer (Dolby etc needs a buffer). So this experimental mod is more for audiophile people.

Differences between Original Thread and the GitHub Repo:
While the original thread on TechPowerUp provides all-in-one packages, the user shibajee extracted the effects and created for each individual brand/effect its own repository, this makes it easier to get your sound effect for your specific brand working. He also provides instructions so you can download the latest Realtek driver and then follow the instructions, which means you do not have to wait for an AiO package update in order to get it working on the latest driver.