Gaming Hardware Buyer's Guide - GPU chart updated
GPU chart updated for Febr. 2020.

Warning/Info:
GamersNexus (see here: https://www.youtube.com/watch?v=IwczmQNHVfo) says that the XFX Radeon RX 5700 XT THICCII is a "bullshit" card (design, cooler failures) however (do not quote me that I wrote bullshit, it's my interpretation of "a awful card), you can contact them to get a replacement for the "cooler problem". I linked to the Reddit thread on purpose, to get background information what happened and how to replace it. The card itself is for the price okay and the reason why it's still listed in the chart (until I found a better replacement)