Workaround for Game Launcher + Games showing "Dev Error 6071"
**Overview:**
In some games like COD: Warzone you might encounter a strange error like e.g. Dev Error 6071. It's unclear what it exactly means and it's unclear if the anti-cheat mechanism is triggering it or not.

However, changing the Machine ID will solve the issue and you can launch the Game or/and the Game launcher. 

There exist a bunch of other games with the same error btw.


**Is it dangerous or will it trigger an anti-cheat system and result in a kick or even perma ban?**
The answer is now, I do not advise you to use the registry tweak unless you suffer from the mentioned error. You MUST also manually backup your existent key first (the ID is unique) BEFORE you apply the tweak. If you have no backup, just delete the entire entry and reboot your machine, which will (if nothing goes wrong) create the entry with a valid ID.


I hope this helps people with this specific error.